+++ src/CheckIn.js
@@ -1,7 +1,4 @@
-import {
-  get,
-  compact,
-} from 'lodash';
+import { get } from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Field, reduxForm } from 'redux-form';
@@ -31,6 +28,7 @@ import {
   Tooltip,
   Dropdown,
 } from '@folio/stripes/components';
+import getEffectiveCallNumber from '@folio/stripes-util/lib/effectiveCallNumber';
 
 import PrintButton from './components/PrintButton';
 import { convertToSlipData } from './util';
@@ -212,6 +210,7 @@ class CheckIn extends React.Component {
         >
           {loan.nextRequest &&
             <PrintButton
+              role="menuitem"
               data-test-print-hold-slip
               buttonStyle="dropdownItem"
               template={this.getTemplate('hold')}
@@ -221,6 +220,7 @@ class CheckIn extends React.Component {
             </PrintButton>}
           {loan.transitItem &&
             <PrintButton
+              role="menuitem"
               data-test-print-transit-slip
               buttonStyle="dropdownItem"
               template={this.getTemplate('transit')}
@@ -231,6 +231,7 @@ class CheckIn extends React.Component {
           {loan.userId &&
             <div data-test-loan-details>
               <Button
+                role="menuitem"
                 buttonStyle="dropdownItem"
                 href={`/users/view/${loan.userId}?layer=loan&loan=${loan.id}`}
                 onClick={(e) => this.handleOptionsChange({ loan, action: 'showLoanDetails' }, e)}
@@ -241,6 +242,7 @@ class CheckIn extends React.Component {
           {loan.userId &&
             <div data-test-patron-details>
               <Button
+                role="menuitem"
                 buttonStyle="dropdownItem"
                 href={`/users/view/${loan.userId}`}
                 onClick={(e) => this.handleOptionsChange({ loan, action: 'showPatronDetails' }, e)}
@@ -250,6 +252,7 @@ class CheckIn extends React.Component {
             </div>}
           <div data-test-item-details>
             <Button
+              role="menuitem"
               buttonStyle="dropdownItem"
               href={`/inventory/view/${loan.item.instanceId}/${loan.item.holdingsRecordId}/${loan.item.id}`}
               onClick={(e) => this.handleOptionsChange({ loan, action: 'showItemDetails' }, e)}
@@ -259,6 +262,7 @@ class CheckIn extends React.Component {
           </div>
           {loan.userId &&
             <Button
+              role="menuitem"
               buttonStyle="dropdownItem"
               href={`/users/view/${loan.userId}`}
               onClick={(e) => this.handleOptionsChange({ loan, action: 'newFeeFine' }, e)}
@@ -268,6 +272,7 @@ class CheckIn extends React.Component {
           {checkinNotePresent &&
             <div data-test-checkin-notes>
               <Button
+                role="menuitem"
                 buttonStyle="dropdownItem"
                 onClick={(e) => this.handleOptionsChange({ loan, action: 'showCheckinNotes' }, e)}
               >
@@ -329,11 +334,7 @@ class CheckIn extends React.Component {
         const inTransitSp = get(loan, ['item', 'inTransitDestinationServicePoint', 'name']);
         return (inTransitSp) ? `${status} - ${inTransitSp}` : status;
       },
-      'effectiveCallNumber': loan => {
-        const callNumberComponents = get(loan, ['item', 'callNumberComponents'], {});
-
-        return compact([callNumberComponents.prefix, callNumberComponents.callNumber, callNumberComponents.suffix]).join(' ');
-      },
+      'effectiveCallNumber': loan => getEffectiveCallNumber(loan),
       ' ': loan => this.renderActions(loan),
     };
 
@@ -370,7 +371,7 @@ class CheckIn extends React.Component {
                         name="item.barcode"
                         validationEnabled={false}
                         placeholder={scanBarcodeMsg}
-                        aria-label={itemIdLabel}
+                        ariaLabel={itemIdLabel}
                         inputRef={barcodeRef}
                         fullWidth
                         component={TextField}